Move ExtractFile from hyperkit to drivers
So that it can be used in other drivers too
